Made some changes to try to speed up compilation
