+++ lib/util.h
@@ -38,20 +38,9 @@ template <typename Iter> Iter select_randomly(Iter start, Iter end) {
     return select_randomly(start, end, gen);
 }
 
-int sample_method(std::vector<int> mds, std::vector<double> wts, int seed) {
-    std::mt19937 gen(seed);
-    std::discrete_distribution<int> dist(wts.begin(), wts.end());
-    int s = dist(gen);
-    return mds[s];
-}
+int sample_method(std::vector<int> mds, std::vector<double> wts, int seed);
 
-int sample_method(std::vector<int> mds, std::vector<double> wts) {
-    std::random_device rd;
-    std::mt19937 gen(rd());
-    std::discrete_distribution<int> dist(wts.begin(), wts.end());
-    int s = dist(gen);
-    return mds[s];
-}
+int sample_method(std::vector<int> mds, std::vector<double> wts);
 
 // Helpers for std::visit
 template <class... Ts> struct overloaded : Ts... { using Ts::operator()...; };
